@@ -1,17 +1,14 @@
-const {stripIndent} = require('common-tags');
 const axios = require('axios');
+const {stripIndent} = require('common-tags');
 
 module.exports = (clients) => {
 	const {rtmClient: rtm, webClient: slack} = clients;
 
 	const notify = async ({type, channel, user}) => {
-		await axios.post('https://slack.com/api/channels.invite', {
-			channel: channel,
-			user: process.env.USER_TSGBOT,
-		}, {
-			headers: {
-				Authorization: `Bearer ${process.env.HAKATASHI_TOKEN}`,
-			},
+		await slack.conversations.invite({
+			channel,
+			users: process.env.USER_TSGBOT, // A comma separated list of user IDs
+			token: process.env.HAKATASHI_TOKEN,
 		});
 
 		const verb = type === 'create' ? '作成' : 'アーカイブから復元';
