@@ -268,7 +268,8 @@ export default ({
                 答えは \`${state.answer
                   .map((dig: number) => String(dig))
                   .join('')}\` だよ:muscle:
-                経過時間: ${round(passedTime / 1000, 2).toFixed(2)}秒`,
+                手数: ${state.history.length}手
+                経過時間: ${round(passedTime / 1000, 3).toFixed(3)}秒`,
                 channel: process.env.CHANNEL_SANDBOX as string,
                 username: 'Hit & Blow',
                 icon_emoji: '1234',
