@@ -9,7 +9,7 @@ const {promisify} = require('util');
 const {escapeRegExp} = require('lodash');
 const {katakanize} = require('japanese');
 
-const loadingPromise = (async () => {
+const loadingPromise = async () => {
 	const englishDictPath = path.resolve(__dirname, 'bep-ss-2.3', 'bep-eng.dic');
 
 	const englishDictExists = await new Promise((resolve) => {
@@ -71,10 +71,10 @@ const loadingPromise = (async () => {
 	return {
 		englishDict,
 	};
-})();
+};
 
 const kanizeEnglish = async (word_) => {
-	const {englishDict} = await loadingPromise;
+	const {englishDict} = await loadingPromise();
 
 	const genLattice = (word) => {
 		// eslint-disable-next-line array-plural/array-plural
