@@ -6,7 +6,7 @@ const path = require("path");
 const {sample, get} = require("lodash");
 const {hiraganize} = require("japanese");
 const {stripIndents} = require("common-tags");
-const {unlock} = require("../achievements");
+const {unlock, increment} = require("../achievements");
 const logger = require('../lib/logger.js');
 
 const stripRe = /^[、。？！,.，．…・?!：；:;\s]+|[、。？！,.，．…・?!：；:;\s]+$/g;
@@ -132,14 +132,19 @@ async function getDictionary() {
 module.exports = (clients) => {
 	const { rtmClient: rtm, webClient: slack } = clients;
 
-	function postMessage(message, channel, broadcast = false) {
+	function postMessage(message, channel, postThreadOptions = {}) {
+		const {broadcast, threadPosted} = {
+			broadcast: false,
+			threadPosted: null,
+			...postThreadOptions,
+		}
 		return slack.chat.postMessage({
 			channel,
 			text: message,
 			as_user: false,
 			username: "pocky",
 			icon_emoji: ":google:",
-			thread_ts: thread,
+			thread_ts: threadPosted ? threadPosted : null,
 			reply_broadcast: broadcast,
 		});
 	}
@@ -175,25 +180,29 @@ module.exports = (clients) => {
 
 		const message = await postMessage(stripIndents`
 			ポッキーゲームを始めるよ～
-			下の単語の〇〇に共通して入る単語は何かな～？
-			スレッドで回答してね!
-
-			${hints.map((hint) => hint.replace(theme.word, '• 〇〇')).join('\n')}
-		`, process.env.CHANNEL_SANDBOX);
+			${hints.map((hint) => hint.replaceAll(theme.word, '〇〇')).join(' / ')}
+		`, process.env.CHANNEL_SANDBOX, {broadcast: false});
 
 		thread = message.ts;
 
-		await postMessage('3分経過で答えを発表するよ～', process.env.CHANNEL_SANDBOX);
+		await postMessage(stripIndents`
+			下の単語の〇〇に共通して入る単語は何かな～？
+			スレッドで回答してね!
+			3分経過で答えを発表するよ～
+
+			${hints.map((hint) => `• ${hint.replaceAll(theme.word, '〇〇')}`).join('\n')}
+		`, process.env.CHANNEL_SANDBOX, {broadcast: false, threadPosted: thread});
+
 		const currentTheme = theme;
 		setTimeout(async () => {
 			if (theme === currentTheme) {
 				await postMessage(stripIndents`
 					なんでわからないの？
 					答えは＊${theme.word}＊ (${theme.ruby}) だよ:anger:
-				`, process.env.CHANNEL_SANDBOX, true);
+				`, process.env.CHANNEL_SANDBOX, {broadcast: true, threadPosted: thread});
 				await postMessage(stripIndents`
 					${hints.map((hint) => hint.replace(theme.word, `• ＊${theme.word}＊`)).join('\n')}
-				`, process.env.CHANNEL_SANDBOX);
+				`, process.env.CHANNEL_SANDBOX, {broadcast: false, threadPosted: thread});
 				theme = null;
 				thread = null;
 			}
@@ -207,14 +216,26 @@ module.exports = (clients) => {
 		const { channel, text, thread_ts, ts } = message;
 		if (theme !== null && thread_ts === thread) {
 			if (text === theme.word || hiraganize(text) === hiraganize(theme.ruby)) {
+				const {word, ruby} = theme;
+				theme = null;
+
 				await postMessage(stripIndents`
 					<@${message.user}> 正解:tada:
-					答えは＊${theme.word}＊ (${theme.ruby}) だよ:tada:
-				`, channel, true);
+					答えは＊${word}＊ (${ruby}) だよ:tada:
+				`, channel, {broadcast: true, threadPosted: thread});
 				await postMessage(stripIndents`
-					${hints.map((hint) => hint.replace(theme.word, `• ＊${theme.word}＊`)).join('\n')}
-				`, channel);
-				theme = null;
+					${hints.map((hint) => hint.replace(word, `• ＊${word}＊`)).join('\n')}
+				`, channel, {broadcast: false, threadPosted: thread});
+				increment(message.user, "pockygame-win");
+				const date = new Date().toLocaleString('en-US', {
+					timeZone: 'Asia/Tokyo',
+					month: 'numeric',
+					day: 'numeric',
+				});
+				if (date === '11/11') {
+					unlock(message.user, "pockygame-on-nov-11");
+				}
+
 				thread = null;
 				return;
 			} else {
@@ -239,10 +260,18 @@ module.exports = (clients) => {
 		}
 		const result = await reply(match[1], match[2].length - 1);
 		if (result !== null) {
-			postMessage(htmlEscape(result), channel);
+			postMessage(htmlEscape(result), channel, {broadcast: false, threadPosted: thread_ts});
 			unlock(message.user, "pocky");
 			if (Array.from(result).length >= 20) {
 				unlock(message.user, "long-pocky");
+			}
+			const date = new Date().toLocaleString('en-US', {
+				timeZone: 'Asia/Tokyo',
+				month: 'numeric',
+				day: 'numeric',
+			});
+			if (date === '11/11') {
+				unlock(message.user, "pocky-on-nov-11");
 			}
 		}
 	});
