@@ -4,7 +4,7 @@ const {promisify} = require('util');
 const schedule = require('node-schedule');
 const {sortBy} = require('lodash');
 const moment = require('moment');
-const {unlock} = require('../achievements');
+const {unlock, increment} = require('../achievements');
 
 class Counter {
 	constructor(name) {
@@ -107,6 +107,9 @@ module.exports = (clients) => {
 	const suspendCounter = new Counter('suspend');
 	const dailyAsaCounter = new Counter('dailyAsa');
 	const weeklyAsaCounter = new Counter('asa');
+	const dailyexerciseCounter = new Counter('dailyexercise');
+	const exerciseCounter = new Counter('exercise');
+	const kasuCounter = new Counter('kasu');
 
 	rtm.on('message', async (message) => {
 		const { channel, text, user, ts: timestamp } = message;
@@ -158,6 +161,29 @@ module.exports = (clients) => {
 				const scores = Array.from(total.entries()).sort(([u1, s1], [u2, s2]) => s2 - s1);
 				const index = scores.findIndex(([u, _]) => u === user);
 				postDM(`あなたの起床点数は${scores[index][1]}点、現在の順位は${index + 1}位`);
+			}
+
+			if (tokens[0] === 'エクササイズランキング' && tokens[1] === '確認') {
+				const total = new Map(exerciseCounter.entries());
+				dailyexerciseCounter.entries().map(([user, score]) => {
+					if (!total.has(user)) {
+						total.set(user, 0);
+					}
+					total.set(user, score + total.get(user));
+				});
+				const scores = Array.from(total.entries()).sort(([u1, s1], [u2, s2]) => s2 - s1);
+				const index = scores.findIndex(([u, _]) => u === user);
+				postDM(`あなたのエクササイズ日数は${scores[index][1]}日、現在の順位は${index + 1}位`);
+			}
+
+			if (tokens[0] === 'カスランキング' && tokens[1] === '確認') {
+				let currentRank = 1;
+				for (let entry of kasuCounter.entries()) {
+					if (entry[0] === user) {
+						return postDM(`あなたのカス数は${entry[1]}回、現在の順位は${currentRank}位`);
+					}
+					currentRank++;
+				}
 			}
 		}
 
@@ -299,6 +325,31 @@ module.exports = (clients) => {
 				dailyAsaCounter.max(user, best_score);
 			}
 		}
+
+		{
+			const kasu = 'カス';
+			if (channel === process.env.CHANNEL_SANDBOX && text.includes(kasu)) {
+				if (user === 'U01AM9D17RV') { // ishitatsuyuki
+					slack.chat.delete({channel, ts: timestamp, token: process.env.HAKATASHI_TOKEN, as_user: true});
+				} else {
+					slack.reactions.add({name: 'kasukasu_dance', channel, timestamp});
+					kasuCounter.add(user);
+				}
+			}
+		}
+
+		{
+			if(text.includes(":exercise-done:")||text.includes(":kintore_houkoku:")){
+				slack.reactions.add({name: 'erai', channel, timestamp})
+				slack.reactions.add({name: 'sugoi', channel, timestamp})
+
+				if (channel.startsWith('C')) {
+					unlock(user, 'first-exercise');
+
+					dailyexerciseCounter.add(user, 1);
+				}
+			}
+		}
 	});
 
 	schedule.scheduleJob('0 19 * * *', async (date) => {
@@ -306,6 +357,11 @@ module.exports = (clients) => {
 			weeklyAsaCounter.add(user, score);
 		});
 		dailyAsaCounter.clear();
+		dailyexerciseCounter.entries().map(([user, score]) => {
+			exerciseCounter.add(user, 1);
+			increment(user, 'exercise-cumulative');
+		});
+		dailyexerciseCounter.clear();
 
 		// on Sundays
 		if (date.getDay() === 0) {
@@ -370,8 +426,14 @@ module.exports = (clients) => {
 					if (index === 0 && weeklyAsaCounter.entries().filter(([, c]) => c === count).length === 1) {
 						unlock(user, 'asa-master');
 					}
+					if (count <= 0) {
+						unlock(user, 'asa-week-0');
+					}
 					if (count >= 720) {
 						unlock(user, 'asa-week-720');
+					}
+					if (count >= 7 * 108) {
+						unlock(user, 'asa-week-perfect');
 					}
 
 					return {
@@ -382,6 +444,51 @@ module.exports = (clients) => {
 			});
 
 			weeklyAsaCounter.clear();
+
+			await slack.chat.postMessage({
+				channel: process.env.CHANNEL_SANDBOX,
+				username: 'exercise-bot',
+				text: '今週のエクササイズランキング',
+				icon_emoji: ':muscle:',
+				attachments: exerciseCounter.entries().map(([user, count], index) => {
+					const member = members.find(({id}) => id === user);
+					if (!member) {
+						return null;
+					}
+					const name = member.profile.display_name || member.name;
+					if (count === 7) {
+						unlock(user, 'everyday-exercise-week');
+					}
+
+					return {
+						author_name: `${index + 1}位: ${name} (${count}日)`,
+						author_icon: member.profile.image_24,
+					};
+				}).filter((attachment) => attachment !== null),
+			});
+
+			exerciseCounter.clear();
+
+			await slack.chat.postMessage({
+				channel: process.env.CHANNEL_SANDBOX,
+				username: 'kasu',
+				text: '今週のカスランキング',
+				icon_emoji: ':kasu:',
+				attachments: kasuCounter.entries().map(([user, count], index) => {
+					const member = members.find(({id}) => id === user);
+					if (!member) {
+						return null;
+					}
+					const name = member.profile.display_name || member.name;
+
+					return {
+						author_name: `${index + 1}位: ${name} (${count}回)`,
+						author_icon: member.profile.image_24,
+					};
+				}).filter((attachment) => attachment !== null),
+			});
+
+			kasuCounter.clear();
 		}
 	});
 }
